(ASC-219) Manage test failure exit code
This commit alters the `execute_tests.sh` script to manage failures
with by tracking the non-zero exit codes. This enables all test steps
of the script to get executed, but will return the exit code of the last
failing test suite.

The setup portion of the script is set to fail immediately, while the
test execution will iteratively complete even when one of the suites
fails. The failing exit code of any process is captured in a global `RC`
variable and is returned as the script's exit code if it is a non-zero
value.