ASC-318 Validate SYS_INVENTORY path
The default path set for `SYS_INVENTORY` recently failed to be the
correct location for the `dynamic_inventory.py` file. This commit
verifies the existence of the file for the set `SYS_INVENTORY` value.
If it fails, it sets it to a known alternative. If the alternative
fails, it is set to an "unknown" location to assist with debugging.