ASC-396 Update test repo name logic
Prior to this commit the `MOLECULE_TEST_REPO` was inheriting the parent
directory of the repo clone path.

In order to make checking the repo name more robust, this commit uses
the `git remote` value to introspect the repo name rather than relying
on the local path name of the clone.