ASC-635 Ensure openstack-ops runs last
This comment ensures that if the `molecule-openstack-ops` molecule is
found in scope for testing, that it is executed last.

Currently, the `molecule-openstack-ops` molecule is run first and
appears to be causing a conflict with the `openstack-ansible-ops`
ansible task for creating nova flavors. Any nova flavors created by the
`molecule-openstack-ops` deployment subsequently fail to deploy when
done so by the `openstack-ansible-ops` task.