ASC-664 Add debug logging to molecule verify
This commit alters the execution of `molecule verify` to include the
`--debug` flag. This outputs detailed debug stack traces in the event of
a failure and writes a debug log to disk. This commit also gathers the
debug logs into the test_results.tar file with the JUnit.xml test result
content.