MAINT - Revert SHAs to match queens on 2018-07-26
The stability of CI jobs has deteriorated. All other branches have been
rolled back to their state on 2018-07-26. This commit pins all of the
molecule submodules to match those of the queens branch as of 2018-07-26
to get to a previously known stable state. The submodule SHAs should
then be carefully updated and validated with the staging CI job against
rpc-o queens branch to reduce the pending molecule updates to only those
that need further work to run successfully in CI.