ASC-877 Skip tests if converge fails
This commit updates the `execute_tests.sh` script to skip running the
`molecule verify` command against a molecule submodule if the preceding
`molecule converge` failed to complete as expected. Prior to this
commit, the tests were executed regardless of the outcome of the
converge setup which caused unexpected test failures due to the testing
environment not being set up as defined in the converge playbook.