Prevent pip version check from outputting stderr
When doing the pip version check, the stderr output if
pip does not exist in the virtualenv is a bit confusing,
and not useful for the test. Rather than allowing it to
leak, we redirect it to /dev/null.

(cherry picked from commit 8bb597d4613b3bf99d3679f755c32fc748bcbec0)