ASC-1029 Use MNAIO ansible inventory for OSP
This commit alters the `execute_tests.sh` script to use the ansible
inventory used for the MNAIO deployment in lieu of the
`dynamic_inventory.json` file that it would use on an RPC-O environment.
Since the moleculerize tool takes a `JSON` file as input, the `INI`
inventory file is converted to a `JSON` file and stored as
`dynamic_inventory.json` as the rest of the script expects.