ASC-1611 Fix argument arrays
This commit updates the arguments that can receive multiple values to
ensure that they append to an array rather than concatenate strings.

The evaluation of the content of the array has also been updated to use
the size of the array rather than retrieving the first value in the
array.